requirements.txt
python-telegram-bot
tenacity==8.2.1
httpx